[ref] Move stock handler views to stock app
