+++ incubator/urls.py
@@ -16,6 +16,7 @@ urlpatterns = patterns(
     url(r'^projects/', include('projects.urls')),
     url(r'^accounts/', include('users.urls')),
     url(r'^space/', include('space.urls')),
+    url(r'^stock/', include('stock.urls')),
 
     url(r'^sm$', 'events.views.sm'),
     url(r'^linux$', 'events.views.linux'),
