+++ users/urls.py
@@ -2,7 +2,7 @@ from django.conf.urls import patterns, url
 from django.contrib.auth.decorators import login_required
 from .views import (
     balance, UserDetailView, UserEditView, spend, top, transfer, show_pamela,
-    hide_pamela, userdetail, product_infos, buy_product_with_stock_handler
+    hide_pamela, userdetail,
 )
 
 urlpatterns = patterns(
@@ -16,7 +16,4 @@ urlpatterns = patterns(
     url(r'^show_pamela$', login_required(show_pamela), name='show_pamela'),
     url(r'^hide_pamela$', login_required(hide_pamela), name='hide_pamela'),
     url(r'^(?P<slug>.+)$', UserDetailView.as_view(), name='user_profile'),
-
-    url(r'^product/(?P<product_barcode>.+)$', product_infos, name='product_infos'),
-    url(r'^stock_handler$', buy_product_with_stock_handler, name='stock_handler_endpoint'),
 )
