+++ users/views.py
@@ -2,16 +2,14 @@ from rest_framework import viewsets
 # from django.core.urlresolvers import reverse
 # from django.http import HttpResponseRedirect
 from django.shortcuts import render, get_object_or_404
-from django.http import HttpResponseRedirect, JsonResponse, HttpResponse
+from django.http import HttpResponseRedirect, JsonResponse
 from django.views.generic.detail import DetailView
 from django.views.generic import UpdateView
 from django.core.urlresolvers import reverse
 from django.contrib import messages
-from django.conf import settings
 from actstream import action
 from actstream.models import Action
-from space.djredis import get_redis, space_is_open
-from space.decorators import private_api, validate
+from space.djredis import get_redis
 
 
 from .serializers import UserSerializer
@@ -28,32 +26,6 @@ def balance(request):
         'topForm': TopForm(),
         'spendForm': SpendForm(),
         'transferForm': TransferForm(),
-    })
-
-
-@private_api(user_qrcode=str, product_barcode=str, quantity=validate(int, lambda x: x > 0))
-def buy_product_with_stock_handler(request, user_qrcode, product_barcode, quantity):
-    if quantity <= 0:
-        return HttpResponseBadRequest("Only positive quantities")
-    prod = get_object_or_404(Product, barcode=product_barcode)
-    user = get_object_or_404(User, qrcode=user_qrcode)
-    user.balance -= quantity * prod.price
-    user.save()
-    action.send(
-        user,
-        verb="a acheté {} pour {}€".format(prod.name, prod.price),
-        public=False
-    )
-    return HttpResponse("ok")
-
-
-def product_infos(request, product_barcode):
-    # => {'price', 'name'}
-    prod = get_object_or_404(Product, barcode=product_barcode)
-    return JsonResponse({
-        'name': prod.name,
-        'category': prod.category.name,
-        'price': prod.price,
     })
